requirements.txt
opentelemetry.api
opentelemetry.sdk
opentelemetry.exporter.otlp.proto.grpc
opentelemetry.instrumentation.logging
pyrfc3339
ghapi
requests
python-dateutil
python-dotenv
